46: Fix the About preview in the CMS
